Changes to remove_padding and moved to utils function
